Update from base

+++ lib/zold/commands/routines/reconnect.rb
@@ -46,7 +46,8 @@ module Zold
         args = ['remote', "--network=#{@opts['network']}"]
         score = @farm.best[0]
         args << "--ignore-node=#{score.host}:#{score.port}" if score
-        cmd.run(args + ['add', 'b1.zold.io', '80']) unless @opts['routine-immediately']
+        cmd.run(args + ['defaults']) unless @opts['routine-immediately']
+        return if @opts['routine-immediately'] && @remotes.all.empty?
         cmd.run(args + ['trim'])
         cmd.run(args + ['select'])
         cmd.run(args + ['update'] + (@opts['never-reboot'] ? [] : ['--reboot']))
